+++ src/App.css
@@ -1,4 +1,4 @@
-/* @import url("https://glitchpop-css.herokuapp.com/library/styles.css"); */
+@import url("https://glitchpop-css.herokuapp.com/library/styles.css");
 @import url("https://fonts.googleapis.com/css2?family=Raleway:wght@400;500;600&display=swap");
 * {
   margin: 0;
