+++ src/Nav/SideBar.js
@@ -78,7 +78,7 @@ function SideBar() {
         </label>
       </div>
       <button
-        className="btn btn-secondary clear-btn"
+        className="btn btn-danger clear-btn"
         type="reset"
         onClick={clearAll}
       >
