+++ src/components/Buttons/CartBtn.js
@@ -15,34 +15,39 @@ const CartBtn = ({ item, isInCart }) => {
         color: "warning",
       });
     }
-    // dispatch({ type: "LOADING", payload: true });
     setLoading(true);
     await addCartItem(item._id);
     try {
       const cartData = await getCartItems();
       dispatch({ type: "SET_CART", payload: cartData });
-      //   dispatch({ type: "LOADING", payload: false });
       setLoading(false);
       throwToast(dispatch, { message: "Added to cart", color: "success" });
     } catch (error) {
       console.log(error);
-      //   dispatch({ type: "LOADING", payload: false });
       setLoading(false);
       throwToast(dispatch, { message: "didn't add to cart", color: "danger" });
     }
   };
-  if (isInCart === undefined) {
+  if (item.stock === "In stock") {
+    if (isInCart === undefined) {
+      return (
+        <button className="card-btn btn-secondary" onClick={addProductToCart}>
+          {loading ? <MiniLoader /> : "Add to cart"}
+        </button>
+      );
+    }
     return (
-      <button className="card-btn btn-secondary" onClick={addProductToCart}>
-        {loading ? <MiniLoader /> : "Add to cart"}
+      <Link to="/cart">
+        <button className="card-btn btn-secondary"> Go to cart</button>
+      </Link>
+    );
+  } else {
+    return (
+      <button className="card-btn" disabled>
+        Add to cart
       </button>
     );
   }
-  return (
-    <Link to="/cart">
-      <button className="card-btn btn-secondary"> Go to cart</button>
-    </Link>
-  );
 };
 
 export default CartBtn;
