+++ src/components/LoaderPage/loaderpage.css
@@ -13,16 +13,16 @@
 
 .lds-dual-ring {
   display: inline-block;
-  width: 15px;
-  height: 15px;
+  width: 25px;
+  height: 25px;
 }
 .lds-dual-ring:after {
   content: " ";
   display: block;
-  width: 15px;
-  height: 15px;
+  width: 25px;
+  height: 25px;
   border-radius: 50%;
-  border: 3px solid rgb(255, 3, 3);
+  border: 2px solid rgb(255, 3, 3);
   border-color: rgb(255, 3, 3) transparent rgb(255, 3, 3) transparent;
   animation: lds-dual-ring 1.2s linear infinite;
 }
