+++ src/components/Product/product.css
@@ -31,6 +31,7 @@
 }
 .card-btn {
   width: 100%;
+  height: 3rem;
 }
 .wishlistBtn {
   width: 200px !important;
